refactor: flatten user feature namespaces
Reorganised domain and persistence classes into Models and Repositories namespaces to simplify the template structure.
