+++ app/Application/Actions/ListUsersAction.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Application\Actions;
 
 use App\Application\DTOs\UserResponseDTO;
-use App\Domain\User\User;
+use App\Domain\Models\User;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
