+++ app/Domain/Models/User.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\User;
+namespace App\Domain\Models;
 
 use JsonSerializable;
