+++ app/Domain/Repositories/UserRepository.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\User;
+namespace App\Domain\Repositories;
+
+use App\Domain\Models\User;
 
 /**
  * Defines the contract for persisting and retrieving user entities.
