+++ tests/Support/Persistence/Repositories/InMemoryUserRepository.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Tests\Support\Persistence\User;
+namespace Tests\Support\Persistence\Repositories;
 
-use App\Domain\User\User;
-use App\Domain\User\UserRepository;
+use App\Domain\Models\User;
+use App\Domain\Repositories\UserRepository;
 
 /**
  * Manages user persistence and retrieval in memory.
